+++ src/Installer/TestSuiteInstaller.php
@@ -155,6 +155,7 @@ class TestSuiteInstaller extends AbstractInstaller
         if ($this->plugin->hasUnitTests()) {
             $this->getOutput()->debug('Build PHPUnit config');
             $processes[] = new MoodleProcess(sprintf('%s/admin/tool/phpunit/cli/util.php --buildconfig', $this->moodle->directory));
+            $processes[] = new MoodleProcess(sprintf('%s/admin/tool/phpunit/cli/util.php --buildcomponentconfigs', $this->moodle->directory));
         }
 
         return $processes;